backend/requirements.txt
fastapi
uvicorn[standard]
autogen-agentchat
autogen-ext[openai]
chromadb
qdrant-client
python-dotenv
pydantic
openai
transformers
sentence-transformers
